Weather History Upload Tool initial upload
The tool is basically complete except for SQL commands. It can load .csv files and parse them, it can connect to the database and has all the basic options I think it needs. It is ready to be setup with the commands to initilize the schema too just to make things easy.